palindrome
palindrome of a no. is refers as when we reverse the number it is equal to original number user had enter. here first take variable named as n , remainder,reverse and temp. now main catch where i commited the mistake was store value of n in temp container i.e temp = n not n = temp . after this apply while loop to n and finally apply if else if temp == reverse. also note that reverse should be initialise with zero